Extract and export functions to parse seriesByTag expressions
